pyproject.toml
[project]
name = "gigaxity-deep-research"
version = "0.1.0"
description = "Deep research MCP server bringing Tongyi DeepResearch 30B to Claude Code via OpenRouter — discover, synthesize, and reason over the web."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Yoloshii" }
]
keywords = [
    "mcp-server",
    "deep-research",
    "model-context-protocol",
    "claude-code",
    "claude-mcp",
    "mcp-tools",
    "agentic-ai",
    "ai-research-agent",
    "ai-research-tool",
    "research-synthesis",
    "agentic-search",
    "openrouter",
    "tongyi-deepresearch",
    "searxng",
    "tavily",
    "gpt-researcher",
    "rag-pipeline",
    "ai-agents",
    "python",
    "fastapi",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tavily-python>=0.3.0",
    "linkup-sdk>=0.2.0",
    "openai>=1.12.0",
    "fastapi-mcp>=0.3.0",
    "fastmcp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yoloshii/gigaxity-deep-research"
Repository = "https://github.com/yoloshii/gigaxity-deep-research"
Issues = "https://github.com/yoloshii/gigaxity-deep-research/issues"
Changelog = "https://github.com/yoloshii/gigaxity-deep-research/releases"
Documentation = "https://github.com/yoloshii/gigaxity-deep-research/blob/main/docs/introduction.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
